Refactor AccountDTO to use BigDecimal for balance and implement user ID retrieval in UserServiceImpl

+++ business-logic/src/main/java/co/com/jdti/businesslogic/dtos/AccountDTO.java
@@ -1,9 +1,11 @@
 package co.com.jdti.businesslogic.dtos;
 
+import java.math.BigDecimal;
+
 public record AccountDTO(
 	String id,
 	String name,
-	double balance,
+	BigDecimal balance,
 	String accountNumber
 ) {
 }
